src/shoppingassistantservice-cloudsql/requirements.txt
flask==3.1.1
langchain-google-genai==2.1.9
langchain==0.3.27
langchain-core==0.3.74
pillow==11.3.0
google-cloud-secret-manager==2.24.0
psycopg2-binary==2.9.9
requests==2.32.5 